refactor: Phase 3 - Split PBReplacerWindow into partial classes
Split the 1000+ line PBReplacerWindow.cs into 4 partial class files:
- PBReplacerWindow.cs: Fields, constants, Unity lifecycle methods (~170 lines)
- PBReplacerWindow.UI.cs: UI initialization methods (~170 lines)
- PBReplacerWindow.Events.cs: Event handlers and command execution (~340 lines)
- PBReplacerWindow.ListView.cs: ListView operations and data handlers (~230 lines)

This improves maintainability by separating concerns and making
each file focused on a specific responsibility.
